Move utils to corresponding file
