+++ static/src/containers/Sounds/utils.js
@@ -1,4 +1,8 @@
-import { MAP_SCALE_FACTOR } from '../../constants';
+import { MAP_SCALE_FACTOR, TSNE_CONFIG, DEFAULT_DESCRIPTOR }
+  from '../../constants';
+import { readObjectPropertyByPropertyAbsName } from '../../utils/objectUtils';
+import tsnejs from '../../vendors/tsne';
+
 
 export const computeSoundGlobalPosition = (tsnePosition, spacePosition, mapPosition) => {
   const { translateX, translateY, scale } = mapPosition;
@@ -7,4 +11,47 @@ export const computeSoundGlobalPosition = (tsnePosition, spacePosition, mapPosit
   const cy = ((tsnePosition.y + (window.innerHeight / (MAP_SCALE_FACTOR * 2))) *
     MAP_SCALE_FACTOR * scale * spacePosition.y) + translateY;
   return { cx, cy };
+};
+
+export const thumbnailMapPosition = { translateX: 0, translateY: 0, scale: 0.3 };
+
+export const getTrainedTsne = (sounds, queryParams) => {
+  const tsne = new tsnejs.Tsne(TSNE_CONFIG);
+  const descriptor = queryParams.descriptor || DEFAULT_DESCRIPTOR;
+  const descriptorKey = `analysis.${descriptor}`;
+  const trainingData = Object.keys(sounds).map(
+    soundID => readObjectPropertyByPropertyAbsName(sounds[soundID], descriptorKey));
+  tsne.initDataRaw(trainingData);
+  return tsne;
+};
+
+const getSoundSpacePosition = (sound, store) => {
+  const { spaces } = store.spaces;
+  // find the space to which sound belongs to
+  const space = spaces.find(curSpace => curSpace.queryID === sound.queryID);
+  return space.position;
+};
+
+export const computePointsPositionInSolution = (tsne, sounds, store) => {
+  const tsneSolution = tsne.getSolution();
+  return Object.keys(sounds).reduce((curState, curSoundID, curIndex) => {
+    const sound = sounds[curSoundID];
+    const mapPosition = store.map;
+    let { spacePosition } = sound;
+    if (!spacePosition) {
+      spacePosition = getSoundSpacePosition(sound, store);
+    }
+    const tsnePosition = {
+      x: tsneSolution[curIndex][0],
+      y: tsneSolution[curIndex][1],
+    };
+    const position = computeSoundGlobalPosition(tsnePosition, spacePosition, mapPosition);
+    return Object.assign({}, curState, {
+      [curSoundID]: {
+        position,
+        spacePosition, // tsne and spacePosition could be useful later (e.g.: when user moves map)
+        tsnePosition,
+      },
+    });
+  }, {});
 };
